move HeaderUtility into source/common/common/matchers and update references
