+++ api/envoy/api/v2/route/BUILD
@@ -9,7 +9,7 @@ api_proto_library_internal(
     deps = [
         "//envoy/api/v2/core:base",
         "//envoy/type:range",
-        "//envoy/type/matcher:header"
+        "//envoy/type/matcher:header",
     ],
 )
