+++ source/common/access_log/BUILD
@@ -45,10 +45,10 @@ envoy_cc_library(
         "//include/envoy/server:access_log_config_interface",
         "//include/envoy/upstream:upstream_interface",
         "//source/common/common:assert_lib",
+        "//source/common/common:matchers_lib",
         "//source/common/common:utility_lib",
         "//source/common/config:utility_lib",
         "//source/common/http:header_map_lib",
-        "//source/common/http:header_utility_lib",
         "//source/common/http:headers_lib",
         "//source/common/http:utility_lib",
         "//source/common/protobuf:utility_lib",
