+++ source/common/config/rds_json.h
@@ -2,9 +2,9 @@
 
 #include "envoy/api/v2/rds.pb.h"
 #include "envoy/api/v2/route/route.pb.h"
-#include "envoy/type/matcher/header.pb.h"
 #include "envoy/json/json_object.h"
 #include "envoy/stats/stats_options.h"
+#include "envoy/type/matcher/header.pb.h"
 
 namespace Envoy {
 namespace Config {
