+++ source/common/router/BUILD
@@ -37,11 +37,11 @@ envoy_cc_library(
         "//source/common/common:assert_lib",
         "//source/common/common:empty_string",
         "//source/common/common:hash_lib",
+        "//source/common/common:matchers_lib",
         "//source/common/common:utility_lib",
         "//source/common/config:metadata_lib",
         "//source/common/config:rds_json_lib",
         "//source/common/config:well_known_names",
-        "//source/common/http:header_utility_lib",
         "//source/common/http:headers_lib",
         "//source/common/http:utility_lib",
         "//source/common/http/websocket:ws_handler_lib",
@@ -188,8 +188,8 @@ envoy_cc_library(
         "//include/envoy/router:router_ratelimit_interface",
         "//source/common/common:assert_lib",
         "//source/common/common:empty_string",
+        "//source/common/common:matchers_lib",
         "//source/common/config:rds_json_lib",
-        "//source/common/http:header_utility_lib",
         "//source/common/protobuf:utility_lib",
     ],
 )
