+++ source/common/router/router_ratelimit.cc
@@ -76,7 +76,7 @@ bool HeaderValueMatchAction::populateDescriptor(const Router::RouteEntry&,
                                                 RateLimit::Descriptor& descriptor,
                                                 const std::string&, const Http::HeaderMap& headers,
                                                 const Network::Address::Instance&) const {
-  if (expect_match_ == Http::HeaderUtility::matchHeaders(headers, action_headers_)) {
+  if (expect_match_ == Matchers::HeaderUtility::matchHeaders(headers, action_headers_)) {
     descriptor.entries_.push_back({"header_match", descriptor_value_});
     return true;
   } else {
