+++ source/extensions/filters/http/fault/BUILD
@@ -25,9 +25,9 @@ envoy_cc_library(
         "//include/envoy/stats:stats_macros",
         "//source/common/common:assert_lib",
         "//source/common/common:empty_string",
+        "//source/common/common:matchers_lib",
         "//source/common/http:codes_lib",
         "//source/common/http:header_map_lib",
-        "//source/common/http:header_utility_lib",
         "//source/common/http:headers_lib",
         "//source/common/protobuf:utility_lib",
         "@envoy_api//envoy/config/filter/http/fault/v2:fault_cc",
