+++ source/extensions/filters/http/health_check/BUILD
@@ -23,9 +23,9 @@ envoy_cc_library(
         "//include/envoy/server:filter_config_interface",
         "//source/common/common:assert_lib",
         "//source/common/common:enum_to_int",
+        "//source/common/common:matchers_lib",
         "//source/common/http:codes_lib",
         "//source/common/http:header_map_lib",
-        "//source/common/http:header_utility_lib",
         "//source/common/http:headers_lib",
         "//source/common/http:utility_lib",
         "//source/common/protobuf:utility_lib",
@@ -39,8 +39,8 @@ envoy_cc_library(
     hdrs = ["config.h"],
     deps = [
         "//include/envoy/registry",
+        "//source/common/common:matchers_lib",
         "//source/common/config:filter_json_lib",
-        "//source/common/http:header_utility_lib",
         "//source/extensions/filters/http:well_known_names",
         "//source/extensions/filters/http/common:factory_base_lib",
         "//source/extensions/filters/http/health_check:health_check_lib",
