+++ test/extensions/filters/http/health_check/health_check_test.cc
@@ -2,7 +2,7 @@
 #include <memory>
 
 #include "common/buffer/buffer_impl.h"
-#include "common/http/header_utility.h"
+#include "common/common/matchers.h"
 #include "common/upstream/upstream_impl.h"
 
 #include "extensions/filters/http/health_check/health_check.h"
@@ -45,7 +45,7 @@ public:
   void prepareFilter(
       bool pass_through,
       ClusterMinHealthyPercentagesConstSharedPtr cluster_min_healthy_percentages = nullptr) {
-    header_data_ = std::make_shared<std::vector<Http::HeaderUtility::HeaderData>>();
+    header_data_ = std::make_shared<std::vector<Matchers::HeaderUtility::HeaderData>>();
     envoy::type::matcher::HeaderMatcher matcher;
     matcher.set_name(":path");
     matcher.set_exact_match("/healthcheck");
